se aplica el metodo de presentacion para usar la clase persona con sus 2 instancias como se pide. Ademas se puede obtener los atributos mediante getters y cambiarlos a gusto mediante setters.
